+++ README.es.md
@@ -1,0 +1,49 @@
+
+
+
+
+# HTML
+  
+Modulo para abrir archivos HTML y extraer o agregar información  
+
+*Read this in other languages: [English](README.md), [Português](README.pr.md), [Español](README.es.md)*
+
+## Como instalar este módulo
+  
+Para instalar el módulo en Rocketbot Studio, se puede hacer de dos formas:
+1. Manual: __Descargar__ el archivo .zip y descomprimirlo en la carpeta modules. El nombre de la carpeta debe ser el mismo al del módulo y dentro debe tener los siguientes archivos y carpetas: \__init__.py, package.json, docs, example y libs. Si tiene abierta la aplicación, refresca el navegador para poder utilizar el nuevo modulo.
+2. Automática: Al ingresar a Rocketbot Studio sobre el margen derecho encontrara la sección de **Addons**, seleccionar **Install Mods**, buscar el modulo deseado y presionar install.  
+
+
+## Overview
+
+
+1. Cargar archivo HTML  
+Carga datos de un archivo HTML en memoria
+
+2. Insertar Etiqueta  
+Inserta una etiqueta en el HTML
+
+3. Guardar archivo  
+Guarda datos en un archivo HTML
+
+4. Cerrar Sesión  
+Quita una sesión de HTML y libera memoria  
+
+
+
+
+----
+### OS
+
+- windows
+- mac
+- linux
+- docker
+
+### Dependencies
+- [**beautifulsoup4**](https://pypi.org/project/beautifulsoup4/)
+### License
+  
+![MIT](https://camo.githubusercontent.com/107590fac8cbd65071396bb4d04040f76cde5bde/687474703a2f2f696d672e736869656c64732e696f2f3a6c6963656e73652d6d69742d626c75652e7376673f7374796c653d666c61742d737175617265)  
+[MIT](http://opensource.org/licenses/mit-license.ph)
