+++ README.md
@@ -1,2 +1,49 @@
+
+
+
+
 # HTML
-Abre archivos .html 
+  
+Module to open HTML files and get or add data  
+
+*Read this in other languages: [English](README.md), [Português](README.pr.md), [Español](README.es.md)*
+
+## How to install this module
+  
+To install the module in Rocketbot Studio, it can be done in two ways:
+1. Manual: __Download__ the .zip file and unzip it in the modules folder. The folder name must be the same as the module and inside it must have the following files and folders: \__init__.py, package.json, docs, example and libs. If you have the application open, refresh your browser to be able to use the new module.
+2. Automatic: When entering Rocketbot Studio on the right margin you will find the **Addons** section, select **Install Mods**, search for the desired module and press install.  
+
+
+## Overview
+
+
+1. Load HTML File  
+Load data from an HTML file into memory
+
+2. Insert Tag  
+Insert tag on HTML
+
+3. Save File  
+Save HTML File
+
+4. End Session  
+Remove an HTML session from memory  
+
+
+
+
+----
+### OS
+
+- windows
+- mac
+- linux
+- docker
+
+### Dependencies
+- [**beautifulsoup4**](https://pypi.org/project/beautifulsoup4/)
+### License
+  
+![MIT](https://camo.githubusercontent.com/107590fac8cbd65071396bb4d04040f76cde5bde/687474703a2f2f696d672e736869656c64732e696f2f3a6c6963656e73652d6d69742d626c75652e7376673f7374796c653d666c61742d737175617265)  
+[MIT](http://opensource.org/licenses/mit-license.ph)
