+++ README.pr.md
@@ -1,0 +1,49 @@
+
+
+
+
+# HTML
+  
+Módulo para abrir arquivos HTML e extrair ou adicionar informações   
+
+*Read this in other languages: [English](README.md), [Português](README.pr.md), [Español](README.es.md)*
+
+## Como instalar este módulo
+  
+Para instalar o módulo no Rocketbot Studio, pode ser feito de duas formas:
+1. Manual: __Baixe__ o arquivo .zip e descompacte-o na pasta módulos. O nome da pasta deve ser o mesmo do módulo e dentro dela devem ter os seguintes arquivos e pastas: \__init__.py, package.json, docs, example e libs. Se você tiver o aplicativo aberto, atualize seu navegador para poder usar o novo módulo.
+2. Automático: Ao entrar no Rocketbot Studio na margem direita você encontrará a seção **Addons**, selecione **Install Mods**, procure o módulo desejado e aperte instalar.  
+
+
+## Overview
+
+
+1. Carregar arquivo HTML  
+Carregar dados de um arquivo HTML na memória
+
+2. Inserir Tag  
+Inserir um tag no HTML
+
+3. Salvar arquivo  
+Salvar dados em um arquivo HTML
+
+4. Terminar Sessão  
+Remove uma sessão HTML e libera memória  
+
+
+
+
+----
+### OS
+
+- windows
+- mac
+- linux
+- docker
+
+### Dependencies
+- [**beautifulsoup4**](https://pypi.org/project/beautifulsoup4/)
+### License
+  
+![MIT](https://camo.githubusercontent.com/107590fac8cbd65071396bb4d04040f76cde5bde/687474703a2f2f696d672e736869656c64732e696f2f3a6c6963656e73652d6d69742d626c75652e7376673f7374796c653d666c61742d737175617265)  
+[MIT](http://opensource.org/licenses/mit-license.ph)
